feat: start dynamic scan on device
