+++ app/components/file-header.js
@@ -14,15 +14,12 @@ const FileHeaderComponent = Ember.Component.extend({
   userRoles: [],
   globalAlpha:0.4,
   radiusRatio:0.9,
-
   isBasicInfo: false,
   isVPNDetails: false,
   isScanDetails: true,
-  apiScanModal: false,
   isLoginDetails: false,
   isOWASPDetails: false,
   isStartingRescan: false,
-  dynamicScanModal: false,
   isRequestingManual: false,
   isDownloadingReport: false,
   showManualScanFormModal: false,
@@ -38,7 +35,6 @@ const FileHeaderComponent = Ember.Component.extend({
   appActions: ENUMS.APP_ACTION.CHOICES.slice(0, -1),
   environments: ENUMS.APP_ENV.CHOICES.slice(0, -1),
 
-  tStartingScan: t("startingScan"),
   tPasswordCopied: t("passwordCopied"),
   tPleaseTryAgain: t("pleaseTryAgain"),
   tManualRequested: t("manualRequested"),
@@ -303,66 +299,6 @@ const FileHeaderComponent = Ember.Component.extend({
       });
     },
 
-    dynamicScan() {
-      const file = this.get("file");
-      file.setBootingStatus();
-      const that = this;
-      const file_id = this.get("file.id");
-      const dynamicUrl = [ENV.endpoints.dynamic, file_id].join('/');
-      this.get("ajax").request(dynamicUrl)
-      .catch(function(error) {
-        file.setNone();
-        that.get("notify").error(error.payload.error);
-      });
-    },
-
-    setAPIScanOption() {
-      const tStartingScan = this.get("tStartingScan");
-      const isApiScanEnabled = this.get("isApiScanEnabled");
-      const project_id = this.get("file.project.id");
-      const apiScanOptions = [ENV.host,ENV.namespace, ENV.endpoints.apiScanOptions, project_id].join('/');
-      const that = this;
-      const data =
-        {isApiScanEnabled};
-      this.get("ajax").post(apiScanOptions, {data})
-      .then(function(){
-        that.get("notify").success(tStartingScan);
-        if(!that.isDestroyed) {
-          that.send("closeModal");
-          that.send("dynamicScan");
-        }
-      })
-      .catch(function(error) {
-        that.get("notify").error(error.payload.error);
-      });
-    },
-
-    doNotRunAPIScan() {
-      triggerAnalytics('feature', ENV.csb.runDynamicScan);
-      this.set("isApiScanEnabled", false);
-      this.send("setAPIScanOption");
-    },
-
-    runAPIScan() {
-      triggerAnalytics('feature', ENV.csb.runAPIScan);
-      this.set("isApiScanEnabled", true);
-      this.send("setAPIScanOption");
-    },
-
-    showURLFilter(param){
-      this.set("showAPIURLFilterScanModal", true);
-      if (param === 'api') {
-        this.set("showAPIScanModal", false);
-        this.set("apiScanModal", true);
-        this.set("dynamicScanModal", false);
-      }
-      if (param === 'dynamic') {
-        this.set("showRunDynamicScanModal", false);
-        this.set("dynamicScanModal", true);
-        this.set("apiScanModal", false);
-      }
-    },
-
     loginRequired() {
       const loginRequiredText = this.$('#app-login-required').val();
       this.set("manualscan.loginRequired", false);
@@ -539,31 +475,6 @@ const FileHeaderComponent = Ember.Component.extend({
       });
     },
 
-    openAPIScanModal() {
-      const platform = this.get("file.project.platform");
-      if ([ENUMS.PLATFORM.ANDROID,ENUMS.PLATFORM.IOS].includes(platform)) { // TEMPIOSDYKEY
-        this.set("showAPIScanModal", true);
-      } else {
-        this.send("doNotRunAPIScan");
-      }
-    },
-
-    goBack() {
-      this.set("showAPIURLFilterScanModal", false);
-      if (this.get("apiScanModal")) {
-        this.set("showAPIScanModal", true);
-      }
-      if (this.get("dynamicScanModal")) {
-        this.set("showRunDynamicScanModal", true);
-      }
-    },
-
-    closeModal() {
-      this.set("showAPIScanModal", false);
-      this.set("showAPIURLFilterScanModal", false);
-      this.set("showRunDynamicScanModal", false);
-    },
-
     closeSubscribeModal() {
       this.set("showSubscribeModal", false);
     },
@@ -588,35 +499,8 @@ const FileHeaderComponent = Ember.Component.extend({
       this.set("showRescanModal", false);
     },
 
-    openRunDynamicScanModal() {
-      this.set("showRunDynamicScanModal", true);
-    },
-
-    closeRunDynamicScanModal() {
-      this.set("showRunDynamicScanModal", false);
-    },
-
     subscribePlan() {
       window.location.href = "/billing";
-    },
-
-    dynamicShutdown() {
-      const file = this.get("file");
-      const that = this;
-      file.setShuttingDown();
-      this.set("isPoppedOut", false);
-      const file_id = this.get("file.id");
-      const shutdownUrl = [ENV.endpoints.dynamicShutdown, file_id].join('/');
-      this.get("ajax").request(shutdownUrl)
-      .then(function() {
-        if(!that.isDestroyed) {
-          file.setNone();
-        }
-      })
-      .catch(function(error) {
-        file.setNone();
-        that.get("notify").error(error.payload.error);
-      });
     },
 
     rescanApp() {
