+++ app/templates/components/dynamic-scan.emblem
@@ -1,0 +1,104 @@
+.column.flex-center
+  div.vnc-box
+    h6.margin-left class=notVNCViewer:hide:show
+      | DYNAMIC SCANNER
+    if file.isNeitherNoneNorReady
+      button.vnc-button.is-primary.margin-top20.is-progress disabled="disabled"
+        .fa-font-size
+          i.fa class="fa-spinner fa-spin"
+        | &nbsp; #{file.statusText}
+    if file.isNoneStatus
+      if file.isDynamicDone
+        button.vnc-button.is-primary.margin-top20 disabled="disabled"
+          = fa-icon "fa-check"
+          | &nbsp; #{t "completed"}
+        if trial.isTrial
+          button.vnc-button.is-primary.margin-top20.margin-left click="openSubscribeModal"
+            .fa-font-size-small
+              = fa-icon "fa-repeat"
+        else
+          button.vnc-button.is-primary.margin-top20.margin-left.ac-dynamic-rescan click="openAPIScanModal"
+            .fa-font-size-small
+              = fa-icon "fa-repeat"
+      else
+        button.vnc-button.is-primary.margin-top20.ac-start-dynamic-scan click="openAPIScanModal"
+          .fa-font-size
+            = fa-icon "fa-play"
+          | &nbsp; #{t "start"}
+    if file.isReady
+      button.is-primary.margin-top20.stop-dynamic-scan click="dynamicShutdown"
+        .fa-font-size-small
+          = fa-icon "fa-stop"
+        |  &nbsp; #{t "stop"}
+
+if notVNCViewer
+  if file.project.isAPIScanEnabled
+    .column.flex-center
+      if file.isRunningApiScan
+        button.is-primary.margin-top20.is-progress disabled="disabled"
+         | #{t "scanning"} : #{file.apiScanProgress}
+         | %
+      else if file.isApiDone
+         button.is-primary.margin-top20 disabled="disabled"
+           = fa-icon "fa-check"
+           | &nbsp; #{t "completed"}
+      else if file.isApiNotDone
+        button.is-primary.margin-top20.ac-start-api-scan click="openRunDynamicScanModal" disabled=file.isNotNoneStatus
+          .fa-font-size
+            = fa-icon "fa-play"
+          | &nbsp; #{t "start"}
+
+= modal-card isActive=showRunDynamicScanModal title=(t "modalCard.runAPIScan.title")
+  .card-body.warning-modal
+    .card-wrapper
+      .api-scanning
+        h6
+          = t "modalCard.runAPIScan.description"
+        form.margin-top20
+          p.control
+          .columns
+            .column
+              button.modal-button.is-primary click="'showURLFilter' 'dynamic'"
+                = t "modalCard.runAPIScan.runScan"
+            .column
+              button.modal-button.is-primary click="closeRunDynamicScanModal"
+                = t "cancel"
+
+= modal-card isActive=showAPIScanModal title=(t "modalCard.apiScan.title")
+  .card-body.warning-modal
+    .card-wrapper
+      .api-scanning
+        h6
+          = t "modalCard.apiScan.description"
+        form.margin-top20
+          p.control
+          .columns
+            .column.flex-center
+              button.modal-button.is-primary.mp-run-api-scan click="'showURLFilter' 'api'"
+                = t "modalCard.apiScan.runAPIScan"
+            .column.flex-center
+              button.modal-button.is-primary.mp-dont-run-api-scan click="doNotRunAPIScan"
+                = t "modalCard.apiScan.dynamicScan"
+
+= modal-card isActive=showAPIURLFilterScanModal title=(t "modalCard.apiScan.title")
+  .card-body.warning-modal
+    .card-wrapper
+      .api-scanning
+        div.warning-band
+          .columns
+            .column.is-one-sixth
+              = fa-icon 'exclamation-triangle'
+            .column
+              = t "modalCard.apiScan.extraDescription"
+        .margin-top20.add-filter-list
+          = api-filter project=file.project
+        div.margin-top20
+          form
+            p.control
+              .columns
+                .column
+                  button.modal-button.is-primary click="runAPIScan"
+                    = t "modalCard.apiScan.startApiScan"
+                .column
+                  button.modal-button.is-primary click="goBack"
+                    =  t "cancel"
